Start making RolesController
Change the name of the column role ( in table users ) to roles_id.
Added home page (only as a structure).
Added bootstrap 4.4.2 in the project.
RolesController made ( resource controller )
Only the index.blade.php view of RolesController is made, the other views for the controller is only made as a blank files.
The routes for RolesResource controller is made.
The todo list is updated.