Update the whole app
Change the layout that all of the files extends.
Add new type of success or error printing.
Make the menu of the app in its final state ( only remains to enter all routes ).
Add an option to all admins to require password update from every user.
Add simple middleware restrictions for the routes.